Fix WebAssembly mysqltest portability and resource cleanup
Correct wasm32 integer formatting, hashing, object alignment, memory allocation, geometry math, and vector-index execution. Add the MySQL byte bridge and focused regressions, release complete SQLite WAL mappings, and actively recycle queued tablets during orderly shutdown.

Validation: 280 cases on actual WebAssembly under Node with WasmFS memory storage: 270 passed, 10 failed, and 0 timed out. Nine residual failures match native; one is a distinct INDEX_LENGTH statistics difference. All 12 engine instances shut down successfully with standard initialization and unchanged resource limits.

+++ src/oblib/lib/alloc/memory_dump.cpp
@@ -740,7 +740,7 @@ void ObMemoryDump::handle(void *task)
       ret = databuff_printf(print_buf_, PRINT_BUF_LEN, print_pos,
           "\n###################%04d-%02d-%02d %02d:%02d:%02d.%06ld###################\n",
           tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday, tm.tm_hour, tm.tm_min,
-          tm.tm_sec, tv.tv_usec);
+          tm.tm_sec, static_cast<int64_t>(tv.tv_usec));
       print_pos += m_task->to_string(print_buf_ + print_pos, PRINT_BUF_LEN - print_pos);
       ret = databuff_printf(print_buf_, PRINT_BUF_LEN, print_pos, "\n");
       // chunk
